refactor: 전화번호 인증요청 로직에 useConfirmState 훅 적용 #204
- 전화번호 인증요청 클릭시 핸들러 함수 외부에서 주입받도록 변경
